refactored whole strategy module into single enum
